generateBriefing/requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
protobuf
google-generativeai
grpcio
google-api-core